+++ ee/apps/ddp-streamer/package.json
@@ -46,6 +46,7 @@
 	},
 	"devDependencies": {
 		"@rocket.chat/eslint-config": "workspace:^",
+		"@swc/core": "^1.3.95",
 		"@types/ejson": "^2.2.1",
 		"@types/gc-stats": "^1.4.2",
 		"@types/meteor": "^2.9.8",
