+++ packages/services-base/package.json
@@ -1,0 +1,35 @@
+{
+	"name": "@rocket.chat/services-base",
+	"version": "0.0.1",
+	"private": true,
+	"devDependencies": {
+		"@types/jest": "~29.5.7",
+		"eslint": "~8.45.0",
+		"jest": "~29.6.4",
+		"ts-jest": "~29.1.1",
+		"typescript": "~5.3.2"
+	},
+	"scripts": {
+		"lint": "eslint --ext .js,.jsx,.ts,.tsx .",
+		"lint:fix": "eslint --ext .js,.jsx,.ts,.tsx . --fix",
+		"test": "jest",
+		"build": "rm -rf dist && tsc -p tsconfig.json",
+		"dev": "tsc -p tsconfig.json --watch --preserveWatchOutput"
+	},
+	"main": "./dist/index.js",
+	"typings": "./dist/index.d.ts",
+	"files": [
+		"/dist"
+	],
+	"dependencies": {
+		"@rocket.chat/core-services": "workspace:^",
+		"@rocket.chat/core-typings": "workspace:^",
+		"@rocket.chat/logger": "workspace:^",
+		"@rocket.chat/message-parser": "workspace:^",
+		"@rocket.chat/models": "workspace:^",
+		"@rocket.chat/ui-contexts": "workspace:^",
+		"eventemitter3": "^4.0.7",
+		"mongodb": "^4.17.2",
+		"pino": "^8.15.0"
+	}
+}
